+++ examples/xor/test.py
@@ -4,7 +4,7 @@ import polars as pl
 
 from flowcean.core.model import Model
 from flowcean.core.tool import test_model
-from flowcean.core.tool.predicates import NotPredicate, PolarsPredicate
+from flowcean.core.tool.testing.predicates import NotPredicate, PolarsPredicate
 from flowcean.polars.environments.dataframe import DataFrame
 from flowcean.polars.transforms.select import Select
