+++ src/flowcean/core/tool/testing/predicates/__init__.py
