+++ src/flowcean/core/tool/testing/predicates/predicate.py
